Factor out non-Material parts of MaterialApp into WidgetsApp
Fixes https://github.com/flutter/flutter/issues/1346
